MagLib: Refactored reading to allow asynchronous reads
Now each mux is set, followed by a function check to return the correct
device object pointer; this is rather than using device1, device2...
Eventually these should actually be moved into an array of objects.

After finding the device, we can then go ahead and do an asynchrnous
request for bytes for each wire1,wire2... This is done with the i2c
sweep variable.

At this point we're just sitting and waiting for all the devices to be
ready, so we just wait for each to have wire.finish() and for bytes to
be avaiable.

Now one by one, go ahead and grab the data as before.