Copyright, which wire, remove magic numbers
Added copyright notice, same as Pete has used so far (MIT) since code is
becoming mature and is publically available.

Now have replaced a lot of the switch-case with a function that returns
the correct wire object. This allows indexing for neatness.

Have replaced some of the 'magic numbers' with constants defined in the
header.